Upload asset with file drop
